Move server.js into src, create middleware defining response local variables
