+++ src/server/middleware.js
@@ -1,0 +1,8 @@
+const setDefaultResponseLocals = (request, response, next) => {
+    response.locals.session = false
+    response.locals.admin = false
+
+  next()
+}
+
+module.exports = setDefaultResponseLocals
